Largest commit in the world :DD

+++ src/main/java/com/sprint/hibernate/repository/ProgressRepository.java
@@ -9,12 +9,12 @@ import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface ProgressRepository extends JpaRepository<Progress, BigInteger> {
+public interface ProgressRepository extends JpaRepository<Progress, Long> {
 
     @Query(value = "SELECT p FROM Progress p JOIN Marathon m WHERE p.trainee.id = :userId AND m.id = :marathonId")
-    List<Progress> allProgressByUserIdAndMarathonId(BigInteger userId, BigInteger marathonId);  //long? UUID?
+    List<Progress> allProgressByUserIdAndMarathonId(long userId, long marathonId);  //long? UUID?
 
     @Query(value = "SELECT p FROM Progress p JOIN Task t WHERE p.trainee.id = :userId AND t.sprint.id = :sprintId")
-    List<Progress> allProgressByUserIdAndSprintId(BigInteger userId, BigInteger sprintId);  // long? UUID?
+    List<Progress> allProgressByUserIdAndSprintId(long userId, long sprintId);  // long? UUID?
 
 }
